BottomNav deluje
